requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
PyYAML==6.0.1
python-dotenv==1.0.1
streamlit==1.34.0
httpx==0.27.0
pandas==2.2.2
pytest==8.2.1
